fix(verus): cast nat args to int in lemma_div_is_ordered call
Verus's lemma_div_is_ordered signature requires int args:
  fn lemma_div_is_ordered(x: int, y: int, z: int)

But `covered * 100` and `total * 100` are `nat * int` (since covered
and total are nat), so the call failed with E0308:

  expected `int`, found `nat`
  at rivet-core/src/verus_specs.rs:249:13 (covered * 100, total * 100)

Surfaced via the bazel test.log artifact upload (PR #209) — without it
the Verus failure was just "FAIL in 0.4s" with no detail.

Cast both products as int explicitly. The third argument `total as int`
was already correctly cast.

Verifies: REQ-004

+++ rivet-core/src/verus_specs.rs
@@ -247,7 +247,7 @@ pub proof fn lemma_coverage_bounded(covered: nat, total: nat)
         // (covered * 100) / total <= (total * 100) / total == 100
         assert((covered * 100) / total <= 100) by {
             vstd::arithmetic::div_mod::lemma_div_is_ordered(
-                covered * 100, total * 100, total as int);
+                (covered * 100) as int, (total * 100) as int, total as int);
             // (total * 100) / total == 100, required to discharge the bound.
             vstd::arithmetic::div_mod::lemma_div_multiples_vanish(
                 100, total as int);
